Adds datasource test (not working)
