--- digitalocean/datasource_digitalocean_firewalls.go
@@ -1,16 +1,0 @@
-package digitalocean
-
-import (
-	"github.com/digitalocean/terraform-provider-digitalocean/internal/datalist"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-)
-
-func dataSourceDigitalOceanFirewalls() *schema.Resource {
-	dataListConfig := &datalist.ResourceConfig{
-		RecordSchema:        firewallSchema(),
-		ResultAttributeName: "firewalls",
-		GetRecords:          getDigitalOceanFirewalls,
-	}
-
-	return datalist.NewResource(dataListConfig)
-}
